chore: move compose to deploy package
