+++ e2e/_suites/helm/helm_charts_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/elastic/e2e-testing/cli/config"
 	"github.com/elastic/e2e-testing/e2e/steps"
 	"github.com/elastic/e2e-testing/internal/common"
-	"github.com/elastic/e2e-testing/internal/compose"
+	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/helm"
 	"github.com/elastic/e2e-testing/internal/kubectl"
 	"github.com/elastic/e2e-testing/internal/shell"
@@ -672,7 +672,7 @@ func InitializeHelmChartTestSuite(ctx *godog.TestSuiteContext) {
 
 		elasticAPMEnvironment := shell.GetEnv("ELASTIC_APM_ENVIRONMENT", "ci")
 		if elasticAPMActive && elasticAPMEnvironment == "local" {
-			serviceManager := compose.NewServiceManager()
+			serviceManager := deploy.NewServiceManager()
 
 			env := map[string]string{
 				"stackVersion": stackVersion,
@@ -726,7 +726,7 @@ func InitializeHelmChartTestSuite(ctx *godog.TestSuiteContext) {
 			}
 
 			if elasticAPMActive {
-				serviceManager := compose.NewServiceManager()
+				serviceManager := deploy.NewServiceManager()
 				err := serviceManager.StopCompose(suiteContext, true, []string{"helm"})
 				if err != nil {
 					log.WithFields(log.Fields{
