+++ e2e/steps/befores.go
@@ -8,14 +8,14 @@ import (
 	"context"
 	"strings"
 
-	"github.com/elastic/e2e-testing/internal/compose"
+	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/shell"
 	log "github.com/sirupsen/logrus"
 )
 
 // AddAPMServicesForInstrumentation adds a Kibana and APM Server instances to the running project
 func AddAPMServicesForInstrumentation(ctx context.Context, profile string, stackVersion string, needsKibana bool, env map[string]string) {
-	serviceManager := compose.NewServiceManager()
+	serviceManager := deploy.NewServiceManager()
 
 	apmServerURL := shell.GetEnv("APM_SERVER_URL", "")
 	if strings.HasPrefix(apmServerURL, "http://localhost") {
